Split classes of internal reader.
- move classes to separate source files
