+++ unit_tests/input/reader_to_storage_test.cpp
@@ -16,7 +16,7 @@
 
 
 #include "input/reader_to_storage.hh"
-#include "input/reader_internal.hh"
+#include "input/reader_internal_base.hh"
 #include "input/accessors.hh"
 
 using namespace std;
